fix: resolve all ESLint errors and warnings (18 → 0)
- AIServicesSection: move all hooks before early return to fix
  react-hooks/rules-of-hooks (7 violations)
- mounts: split registry functions into mounts-registry.ts so mounts.tsx
  only exports MountPoint component (fixes 3 react-refresh violations)
- mounts-registry: convert empty interface declarations to type aliases
  to fix no-empty-object-type (4 violations)
- mounts.test.ts: replace `as any` with typed casts via ComponentType<T>
  to fix no-explicit-any (3 violations)
- vite.config.ts: remove stale no-console eslint-disable directive (1 warning)
- Update all import sites to use mounts-registry for non-component exports
